Throw error if ticket does not exist

+++ MMS-Backend/src/main/java/ca/mcgill/ecse321/MMSBackend/service/TicketService.java
@@ -3,6 +3,8 @@ package ca.mcgill.ecse321.MMSBackend.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.List;
 import ca.mcgill.ecse321.MMSBackend.dao.TicketRepository;
 import ca.mcgill.ecse321.MMSBackend.model.Client;
 import ca.mcgill.ecse321.MMSBackend.model.Ticket;
@@ -21,7 +23,6 @@ public class TicketService {
         ticket.setFee(client.getMuseumManagementSystem().getTicketFee());
         ticket.setClient(client);
         ticket.setMuseumManagementSystem(client.getMuseumManagementSystem());
-        ticket.setTicketId(ticket.getTicketId());
         ticket.setIsActive(true);
         ticketRepository.save(ticket);
         return ticket;
@@ -33,7 +34,15 @@ public class TicketService {
      */
     @Transactional
     public Ticket getTicket(int ticketId) {
-        return ticketRepository.findTicketByTicketId(ticketId);
+
+        Ticket ticket = ticketRepository.findTicketByTicketId(ticketId);
+
+        if(ticket == null) {
+            throw new IllegalArgumentException("Ticket does not exist");
+        }else{
+            return ticket;
+        }
+        
     }
 
     /**
@@ -49,8 +58,14 @@ public class TicketService {
      */
     @Transactional
     public void deleteTicket(int ticketId) {
+
         Ticket ticket = ticketRepository.findTicketByTicketId(ticketId);
-        ticketRepository.delete(ticket);
+
+        if(ticket == null) {
+            throw new IllegalArgumentException("Ticket does not exist");
+        }else{
+            ticketRepository.delete(ticket);
+        }
     }
 
     /**
@@ -58,8 +73,14 @@ public class TicketService {
      */
     @Transactional
     public void setTicketStatus(int ticketId, boolean status) {
+
         Ticket ticket = ticketRepository.findTicketByTicketId(ticketId);
-        ticket.setIsActive(status);
+        
+        if(ticket == null) {
+            throw new IllegalArgumentException("Ticket does not exist");
+        }else{
+            ticket.setIsActive(status);
+        }
     }
 
     /**
@@ -67,8 +88,32 @@ public class TicketService {
      */
     @Transactional
     public void getTicketStatus(int ticketId) {
+
         Ticket ticket = ticketRepository.findTicketByTicketId(ticketId);
-        ticket.getIsActive();
+
+        if(ticket == null) {
+            throw new IllegalArgumentException("Ticket does not exist");
+        }else{
+            ticket.getIsActive();
+        }
+     
+    }
+
+    /**
+     * Get all tickets by client
+     */
+    public List<Ticket> getAllTicketsByClient(Client client){
+        
+        List<Ticket> ticketsByClient = new ArrayList<>();
+        Iterable<Ticket> tickets = ticketRepository.findAll();
+        
+        for(Ticket ticket : tickets){
+            if(ticket.getClient().equals(client)){
+                ticketsByClient.add(ticket);
+            }
+        }
+
+        return ticketsByClient;
     }
 
 }
